move footer file to components folder
